requirements.txt
streamlit
google-generativeai
python-dotenv
langchain_google_genai
langchain